+++ src/main/java/com/capgemini/service/ILoanServiceimpl.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.capgemini.entites.LoanProgram;
+import com.capgemini.exception.LoanNotFoundException;
 import com.capgemini.repository.LoanProgramRepository;
 
 @Service
@@ -20,9 +21,9 @@ public class ILoanServiceimpl implements ILoanService {
 
 	@Override
 	public String updateLoanDetails(int loanid, LoanProgram loanprogram) {
-		// if(!loanprogramrepository.existsById(loanid)) {
-		// throw new LoannotfoundException("Check the id and Try again");
-		// }
+		if (!loanprogramrepository.existsById(loanid)) {
+			throw new LoanNotFoundException("Check the id and Try again");
+		}
 		LoanProgram ld = loanprogramrepository.findById(loanid).get();
 		ld.setLoantype(loanprogram.getLoantype());
 		loanprogramrepository.save(ld);
@@ -33,6 +34,9 @@ public class ILoanServiceimpl implements ILoanService {
 
 	@Override
 	public String deleteLoanDetails(int loanid) {
+		if (!loanprogramrepository.existsById(loanid)) {
+			throw new LoanNotFoundException("The id is not available ");
+		}
 		loanprogramrepository.deleteById(loanid);
 		return "Record deleted";
 	}
